Revert "minor improvements, added routing"
