+++ front_app/package.json
@@ -16,8 +16,6 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-i18next": "^14.1.2",
-    "react-router": "^6.27.0",
-    "react-router-dom": "^6.27.0",
     "swiper": "^11.1.4"
   },
   "devDependencies": {
