+++ front_app/src/i18n/translations/russian.json
@@ -4,8 +4,8 @@
   "header.nav.address": "Адрес",
   "header.nav.form": "Анкета",
 
-  "main.saturday": "суббота",
-  "main.september": "ноября",
+  "main.saturday": "Суббота",
+  "main.september": "Ноябрь",
 
   "about.title": "",
   "about.subtitle": "Дорогие гости, приглашаем <br/> Вас стать дорогим гостем свадьбы <br /> <span>Арнура и Жанерке!</span>",
